% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_modifications}
\alias{calc_modifications}
\alias{calc_A_routine}
\alias{calc_B_routine}
\title{Package functions}
\usage{
calc_modifications(x)

calc_A_routine(x)

calc_B_routine(x)
}
\arguments{
\item{x}{Vector of Numeric Values with length \eqn{n}.}
}
\value{
A vector of numeric values with length \eqn{n}.
}
\description{
Functions listed here are derived from the exported \emph{C++} functions.
Note that all functions work even after being split into multiple separate
files.
}
\examples{
my_vec = c(3.5, 8, -1.5, -2, 0, 42)
calc_modifications(my_vec)
my_vecA = c(881, 884, 1,2,-4)
calc_A_routine(my_vecA)
my_vecB = c(512,32, -1, 1.3, 5.9)
calc_B_routine(my_vecB)
}
