% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SrcDir-package.R
\docType{package}
\name{SrcDir-package}
\alias{SrcDir}
\alias{SrcDir-package}
\title{SrcDir: Embedding Multiple Code Files in Src}
\description{
Provides an implementation showing how multiple compiled code
    files in the `src` directory can be used.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/r-pkg-examples/rcpp-headers-src}
}

}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu}

}
